102-fibonacci.c

+++ functions_nested_loops/102-fibonacci.c
@@ -8,14 +8,14 @@
 int main(void)
 {
 	int i;
-	long long int fib, init = 1, sec = 2;
+	long int fib, init = 1, sec = 2;
 
-	printf("%lld, ", init);
-	printf("%lld", sec);
+	printf("%ld, ", init);
+	printf("%ld", sec);
 	for (i = 0; i < 50; i++)
 	{
 		fib = init + sec;
-		printf(", %lld", fib);
+		printf(", %ld", fib);
 		init = sec;
 		sec = fib;
 	}
